RfidTool 和 UHF PC.UMI 有关的一处 bug 修正

+++ RfidDrivers.First/RfidDriver1.cs
@@ -4338,7 +4338,7 @@ out Reader reader);
                                     continue;
                                 }
 
-                                pc.AFI = enable ? 0x07 : 0xc2;
+                                pc.AFI = enable ? (byte)0x07 : (byte)0xc2;
                                 var new_pc_bytes = UhfUtility.EncodePC(pc);
 
                                 Debug.Assert(new_pc_bytes.Length == 2);
